[DF-10357] bugfix
Former-commit-id: 56b76e638efdd9132ba00f20c52e8c9ec63a1a34

+++ Tools/Launcher/Classes/main.cpp
@@ -15,15 +15,20 @@ int main(int argc, char* argv[])
 
     QFileInfo fi(argv[0]);
     QDir currentDir(fi.absoluteDir());
-    QString platformsDir = "platforms";
-    QString windowsDll = "qwindows.dll";
-    currentDir.cd(platformsDir);
-    currentDir.removeRecursively();
-    currentDir.cdUp();
+    QString platformsPath = "platforms";
+    QString windowsDllPath = "qwindows.dll";
+    //remove "platforms" with permission hack
+    if (currentDir.exists(platformsPath))
+    {
+        currentDir.cd(platformsPath);
+        currentDir.removeRecursively();
+        currentDir.cdUp();
+    }
+    //try copy even if folder already exists
     //if (!currentDir.exists(platformsDir))
     {
-        currentDir.mkpath(platformsDir);
-        QFile::copy(windowsDll, platformsDir + "/" + windowsDll);
+        currentDir.mkpath(platformsPath);
+        QFile::copy(windowsDllPath, platformsPath + "/" + windowsDllPath);
     }
 #endif //windows
     QApplication a(argc, argv);
